requirements.txt
git+git://github.com/avwx-rest/avwx-api-core.git@c176ce165886634c3df7a8927050c9c1973948b7#egg=avwx_api_core

avwx-engine==1.6.3
httpx~=0.16
hypercorn>=0.11
rollbar>=0.15
voluptuous~=0.12